refactor: 홈페이지 카드들 -> cards 로 이관작업
- 헤더에 일반 거래 추가
